app/apis/requirements.lpr.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
ultralytics>=8.0.0
easyocr>=1.6.0
numpy>=1.21.0
opencv-python>=4.5.3
pillow>=8.3.1
requests>=2.26.0
opentelemetry-api>=1.11.1
opentelemetry-sdk>=1.11.1
opentelemetry-exporter-otlp>=1.11.1
opentelemetry-instrumentation-fastapi>=0.30b1
opentelemetry-instrumentation-requests>=0.30b1
prometheus-client>=0.14.1
prometheus-fastapi-instrumentator>=5.9.1